add defaults to builders remove old models
